pyproject.toml
[project]
name = "stonksfeed"
version = "0.1.7"
description = "A stock investor's news feed par excellence!"
readme = "README.md"
requires-python = ">=3.12"
license = "GPL-3.0-or-later"
authors = [
    { name = "m1yag1", email = "mike.arbelaez@gmail.com" }
]

# Runtime dependencies for Lambda (also used by tests)
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "moto[dynamodb]>=5.0.0",
    "boto3>=1.34.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
# Skip DynamoDB integration test by default (requires local DynamoDB)
addopts = "--ignore=tests/unit/test_dynamodb.py"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
